+++ minishell/src/parse_user_input/parse_user_input.c
@@ -6,7 +6,7 @@
 /*   By: jaemoojung <jaemoojung@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/02 20:57:26 by jaemoojung        #+#    #+#             */
-/*   Updated: 2022/03/03 00:41:09 by jaemoojung       ###   ########.fr       */
+/*   Updated: 2022/03/06 22:57:30 by jaemoojung       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,10 +19,10 @@ int	parse_user_input(const char *input, t_ast_node **tree)
 	t_token	*tmp;
 
 	splitted = mvs_split(input);
-	debug_print_vector(splitted);
+	// debug_print_vector(splitted);
 	lexer(splitted, &tokens);
 	tmp = tokens;
-	debug_print_tokens(tmp);
+	// debug_print_tokens(tmp);
 	parser(&tmp, tree);
 	return (0);
 }
