+++ minishell/src/parse_user_input/parser_utils.c
@@ -6,7 +6,7 @@
 /*   By: jaemoojung <jaemoojung@student.42.fr>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/03/06 17:06:32 by jaemoojung        #+#    #+#             */
-/*   Updated: 2022/03/06 17:12:47 by jaemoojung       ###   ########.fr       */
+/*   Updated: 2022/03/06 22:57:13 by jaemoojung       ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
